Front-End-React-App

+++ frontend/src/Screens/SignupScreen.js
@@ -16,8 +16,10 @@ export default function SigninScreen() {
     const redirectInUrl = new URLSearchParams(search).get('redirect');
     const redirect = redirectInUrl ? redirectInUrl : '/';
 
-    const [email, setEmail] = useState(' ');
-    const [password, setPassword] = useState(' ');
+    const [name, setName] = useState('');
+    const [email, setEmail] = useState('');
+    const [password, setPassword] = useState('');
+    const [confirmPassword, setConfirmPassword] = useState('');
 
     const { state, dispatch: ctxDispatch } = useContext(Store);
     const {userInfo} = state;
@@ -25,10 +27,16 @@ export default function SigninScreen() {
 
     const submitHandler = async (e) => {
         e.preventDefault();
+        if (password !== confirmPassword) {
+            toast.error('Passwords do not match');
+            return;
+        }
         try {
-            const { data } = await Axios.post('/api/users/signin', {
+            const { data } = await Axios.post('/api/users/signup', {
+                name,
                 email,
                 password,
+                confirmPassword,
             });
             ctxDispatch({type: 'USER_SIGNIN', payload: data})
             localStorage.setItem('userInfo', JSON.stringify(data));
@@ -46,28 +54,32 @@ export default function SigninScreen() {
     return (
       <Container className="small-container">
         <Helmet>
-            <title>Sign In</title>
+            <title>Sign Up</title>
         </Helmet>
-        <h1 className="my-3">Sign In</h1>
+        <h1 className="my-3">Sign Up</h1>
         <Form onSubmit={submitHandler}>
+            <Form.Group className="mb-3" controlId="name">
+                <Form.Label>Name</Form.Label>
+                <Form.Control  onChange={(e) => setName(e.target.value)}/>
+            </Form.Group>
             <Form.Group className="mb-3" controlId="email">
                 <Form.Label>Email</Form.Label>
-                <Form.Control type='email' required   onChange={(e) => setEmail(e.target.value)}
-
-                />
+                <Form.Control type='email' required   onChange={(e) => setEmail(e.target.value)}/>
             </Form.Group>
             <Form.Group className="mb-3" controlId="password">
                 <Form.Label>Password</Form.Label>
-                <Form.Control type='password' required   onChange={(e) => setPassword(e.target.value)}
-
-                />
+                <Form.Control type='password' required   onChange={(e) => setPassword(e.target.value)} />
+            </Form.Group>
+            <Form.Group className="mb-3" controlId="confirmpassword">
+                <Form.Label>Confirm Password</Form.Label>
+                <Form.Control type='password' required   onChange={(e) => setConfirmPassword(e.target.value)} />
             </Form.Group>
             <div className="mb-3">
-                <Button type="submit">Sign In</Button>
+                <Button type="submit">Sign Up</Button>
             </div>
             <div className="mb-3">
-                New Customer?{' '}
-                <Link to={`/signup?redirect=${redirect}`}>Create your account</Link>
+                Already have an account?{' '}
+                <Link to={`/signin?redirect=${redirect}`}>Sign-In</Link>
             </div>
         </Form>
       </Container>
